fix(ci): make main green — full-feature compile, deny advisories, zero-warning clippy gate
- core: fix 18 compile errors under extended feature set
  (ToolResult consolidation fallout, rusqlite u64 FromSql, AES-GCM
  usize/u64 drift, orphaned provider mapping)
- auth: quick-xml 0.31 -> 0.42 migration, rewrite broken RBAC guard
  (axum extension-based extractor), fix SAML issuer/self-closing
  tag parsing bugs the new tests exposed
- gateway: AdminState roles init, fix missing .await, feature-gated
  stubs, docs
- deps: anyhow/h2/crossbeam-epoch/lru-0.18/quick-xml/wasmtime-46
  bumps; deny.toml ignore list reconciled (advisories now ok)
- clippy: clawdius-unsafe as_chunks (Rust 1.98 lint), UI/tauri/tests
  pedantic fixes; workspace now passes -D warnings

+++ crates/clawdius-core/src/encryption.rs
@@ -316,7 +316,7 @@ fn fill_random(buf: &mut [u8]) {
     }
     // Process remaining bytes with u128 (16-byte chunks)
     while i + 16 <= len {
-        let chunk_idx = i / 8;
+        let chunk_idx = (i / 8) as u64;
         let lo = seed
             .wrapping_mul(chunk_idx.wrapping_add(1))
             .wrapping_add(0x9e3779b97f4a7c15);
@@ -329,7 +329,7 @@ fn fill_random(buf: &mut [u8]) {
     }
     // Process remaining bytes individually
     while i < len {
-        let chunk_idx = i / 8;
+        let chunk_idx = (i / 8) as u64;
         let val = seed
             .wrapping_mul(chunk_idx.wrapping_add(1))
             .wrapping_add(0x9e3779b97f4a7c15);
